♻️ refactor #136 : 초대 리스트 전역 관리 설정 완료 및 파일 변경
